requirements.txt
numpy>=1.24,<2.0
pandas>=2.1,<3.0
scipy>=1.11,<2.0
scikit-learn>=1.3,<2.0
lightgbm>=4.0,<5.0
catboost>=1.2,<2.0
xgboost>=1.7,<2.0
sentence-transformers>=2.2,<3.0
torch>=2.1,<3.0
torchvision>=0.16,<1.0
open-clip-torch>=2.23,<3.0
optuna>=3.4,<4.0
pillow>=10.0,<11.0
requests>=2.31,<3.0
pyarrow>=13.0,<15.0
joblib>=1.3,<2.0
tqdm>=4.66,<5.0
python-dotenv>=1.0,<2.0
huggingface-hub>=0.19,<1.0
ftfy>=6.1,<7.0
regex>=2023.10.0